requirements.txt
pandas
sqlalchemy
psycopg2-binary
google-cloud-storage
google-cloud-bigquery
apache-airflow-providers-google
google-cloud-pubsub
apache-beam[gcp]
apache-airflow-providers-postgres
requests
